+++ serving-core/src/main/scala/com/stratio/sparta/serving/core/constants/AppConstant.scala
@@ -108,6 +108,9 @@ object AppConstant extends ZookeeperUtils {
   //Tenant variable to use as identity
   val spartaTenant = Properties.envOrElse(MarathonConstant.TenantName, "sparta")
 
+  //Multi-tenant sid
+  lazy val spartaServerMarathonAppId = Properties.envOrElse(MarathonConstant.SpartaServerMarathonAppId, s"/sparta/$instanceServiceName/$instanceServiceName")
+
   //Security ON/OFF
   val securityTLSEnable = Try(Properties.envOrElse(MarathonConstant.SpartaTLSEnableEnv, "false").toBoolean).getOrElse(false)